Perform entity operations using command (#313)
Execute commands to implement all entity operations (e.g. add/remove components, set property, etc.). The helps to ensure that operations are performed sequentially, reducing the scope for concurrent execution bugs.

To support this, the new CommandService.ExecuteAsync() method allows the client to enqueue a command and wait for it to execute, returning the execution result.

I've also refactored the command & entity undo/redo APIs to be consistent with each other and easier to use.


* Add CommandService.ExecuteAsync

* AddComponent command

* Delete component using ExecuteAsync

* Remove unnecessary Task.Delay() calls

* Changed MoveComponent and CopyComponent to use commands

* StringField uses SetPropertyCommand

* Execute command for entity undo/redo

* Make CreateCommand and EnqueueCommand private

* Use default parameters instead of overloaded Execute methods.

* Rename ExecuteNow() to ExecuteImmediate()

* Change ExecuteAsync() to allow undoable commands

* Update comments

* Make entity command static methods async