Display Project Name in Application Title Bar, Fix for issue #415
Move the Project title into the application title bar (Currently this is windows only, so ensure the original behaviour is still in place on non-Windows platforms). When this is the case change the Explorer panel title to now read 'Explorer', similarly to the behaviour of the other application panels.

This commit fixes issue #415.